+++ src-tests/Tests.hs
@@ -46,23 +46,23 @@ main = defaultMain $ testGroup "Tests" [unitTests,props]
                   (forAll arbitraryHashOptions $ \hashOptions ->
                    forAll arbitraryPassword $ \password ->
                    forAll arbitrarySalt $ \salt ->
-                         verifyEncoded (hashEncoded hashOptions password salt) password)
+                         verifyEncoded (either undefined id $ hashEncoded hashOptions password salt) password == Argon2Ok)
 
        , testProperty "Unencoded hashing"
                   (forAll arbitraryHashOptions $ \hashOptions ->
                    forAll arbitraryPassword $ \password ->
                    forAll arbitrarySalt $ \salt ->
-                          hash hashOptions password salt /= password)
+                          either undefined id (hash hashOptions password salt) /= password)
 
        , testProperty "defaultHashOptions"
                   (forAll arbitraryVariant $ \variant ->
                   (forAll arbitraryPassword $ \password ->
                    forAll arbitrarySalt $ \salt ->
-                          verifyEncoded (hashEncoded
-                                    (defaultHashOptions {hashVariant = variant})
-                                    password
-                                    salt)
-                                 password))
+                          verifyEncoded (either undefined id $
+                                         hashEncoded (defaultHashOptions {hashVariant = variant})
+                                         password
+                                         salt)
+                                        password == Argon2Ok))
        ]
 
 
@@ -76,16 +76,22 @@ main = defaultMain $ testGroup "Tests" [unitTests,props]
                                        , hashVersion     = Argon2Version13
                                        }
          in [ testCase "hashEncoded" $
-              hashEncoded opts "password" "somesalt" @?= "$argon2i$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG"
+              hashEncoded opts "password" "somesalt" @?= Right "$argon2i$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG"
             , testCase "verifyEncoded" $
-              verifyEncoded "$argon2i$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG" "password" @?= True
+              verifyEncoded "$argon2i$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG" "password" @?= Argon2Ok
             , testCase "verifyEncoded 2" $
-              verifyEncoded "$argon2i$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG" "passvord" @?= False
+              verifyEncoded "$argon2i$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG" "passvord" @?= Argon2VerifyMismatch
             , testCase "verifyEncoded 3" $
-              verifyEncoded "$argon2d$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG" "password" @?= False
+              verifyEncoded "$argon2d$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG" "password" @?= Argon2VerifyMismatch
             , testCase "verifyEncoded 4" $
-              verifyEncoded "$argon2id$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG" "password" @?= False
+              verifyEncoded "$argon2id$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG" "password" @?= Argon2VerifyMismatch
+            , testCase "verifyEncoded 5" $
+              verifyEncoded "$argon2id$v=19$m=1,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG" "password" @?= Argon2MemoryTooLittle
+            , testCase "verifyEncoded 6" $
+              verifyEncoded "$argon2x$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJObG" "password" @?= Argon2DecodingFail
+            , testCase "verifyEncoded 7" $
+              verifyEncoded "$argon2id$v=19$m=65536,t=2,p=4$c29tZXNhbHQ$RdescudvJCsgt3ub+b+dWRWJTmaaJ" "password" @?= Argon2DecodingFail
             , testCase "hash" $
-              hash opts "password" "somesalt" @?= "\x45\xD7\xAC\x72\xE7\x6F\x24\x2B\x20\xB7\x7B\x9B\xF9\xBF\x9D\x59\x15\x89\x4E\x66\x9A\x24\xE6\xC6"
+              hash opts "password" "somesalt" @?= Right "\x45\xD7\xAC\x72\xE7\x6F\x24\x2B\x20\xB7\x7B\x9B\xF9\xBF\x9D\x59\x15\x89\x4E\x66\x9A\x24\xE6\xC6"
             ]
        ]